apps/agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0
boto3>=1.35.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
numpy>=1.26.0
python-dotenv>=1.0.0
mcp>=1.0.0
uvicorn>=0.30.0
fastapi>=0.115.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

